+++ Postfixform/Postfixform/Postfix.h
@@ -1,4 +1,4 @@
 #pragma once
 
 // Function that translates an expression from a postfix form to an infix form
-int countTheExpression(char* postfixEntry);
+float countTheExpression(char* postfixEntry);
